Add writeJSON function to helpers
Implemented a new function 'writeJSON' in helpers.go that takes responsibility for marshaling a Go value into JSON, setting some custom headers, and sending a HTTP response. This function takes an http.ResponseWriter, a HTTP status code, the data to encode to JSON, and a map of header names and values as parameters.