Add Models struct and NewModels constructor in data package
This adds a new file, internal/data/models.go, where we define a new `Models` struct to handle all data models and a `NewModels` constructor that initializes a `Models` instance with a `MovieModel` struct. An error variable, `ErrRecordNotFound`, is also included for handling instances when a record is not found.