Refactor JSON marshalling in writeJSON function
The writeJSON function in helpers.go has been refactored in this commit. The JSON marshalling of data, which was previously at the end of the function, has now been moved to the start. This change makes it easier to handle any errors that might occur during the process right at the beginning.