Implement Get method and error handling in MovieModel
Added the Get method in MovieModel in movies.go, which fetches a movie by id from the database. If the id is less than 1, or if the query does not find a film with the given id, an record not found error is returned. All other errors from the database method are passed along as is.