Add "GetAll" function to MovieModel in movies.go
Implemented the "GetAll" function in the "movies.go" file, which retrieves all movies from the database responding to specific filters. It queries the database, scans the results into a movie object and then appends it to the movies array. Added a context with timeout to control for long running queries.